Add: introduce `updateColoniesNativeTokenStatuses` util
